requirements.txt
Flask==2.3.2
openai==1.2.3
requests==2.31.0
gunicorn==21.2.0
flask-cors==6.0.0
python-dotenv==1.0.0
